Added pem and geoip files

+++ NetProbe/Utility/TestStorage.h
@@ -17,9 +17,9 @@
 
 + (NSArray*)remove_test:(NSNumber*)test_id;
 
-+ (NSArray*)remove_test_atindex:(int)index;
++ (NSArray*)remove_test_atindex:(long)index;
 
-+ (NetworkMeasurement*)get_test_atindex:(int)index;
++ (NetworkMeasurement*)get_test_atindex:(long)index;
 
 + (void)set_completed:(NSNumber*)test_id;
